+++ spec/rspec/core/configuration_spec.rb
@@ -531,17 +531,10 @@ module RSpec::Core
       end
     end
 
-    describe "#debug=true" do
-      it "requires 'ruby-debug'" do
-        config.should_receive(:require).with('ruby-debug')
+    describe "#debug=" do
+      it "is deprecated" do
+        RSpec.should_receive(:warn_deprecation)
         config.debug = true
-      end
-    end
-
-    describe "#debug=false" do
-      it "does not require 'ruby-debug'" do
-        config.should_not_receive(:require).with('ruby-debug')
-        config.debug = false
       end
     end