+++ spec/rspec/core/kernel_extensions_spec.rb
@@ -1,9 +1,12 @@
 require 'spec_helper'
 
 describe "extensions" do
-  describe "debugger" do
-    it "is defined on Kernel" do
-      Kernel.should respond_to(:debugger)
+  describe "#debugger" do
+    it "warns if ruby-debug is not installed" do
+      object = Object.new
+      object.should_receive(:warn).with(/debugger .* ignored/)
+      object.stub(:require) { raise LoadError }
+      object.method_missing(:debugger)
     end
   end
 end